secure-email-service/src/secure-email-service/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
jinja2==3.1.5
redis==5.2.1
cryptography==44.0.0
python-smail==0.9.0
playwright==1.49.1
oscrypto@git+https://github.com/wbond/oscrypto.git@d5f3437